add --color=[auto|never|always] command line option
Adopt the --color flag of ls(1) so that people who wish
to turn off the pretty colors may do so by adding
--color=never to an alias.

The default is auto, which matches the previous default:
use colors when the output is a tty.

One may also now specify --color=always to get color codes sent to
pipes, for example when combined with "less -r".

Signed-off-by: Bob Copeland <copeland@lastpass.com>