edit/generate: allow creating sites under shared folders
If the group name starts with Shared-XXX, then site is
most likely supposed to go in a shared folder.  When a new
site is created, search the blob for shared folders with the
matching name, and save that into account->share so that
updates to the fields will use the shared folder's key.

Signed-off-by: Bob Copeland <copeland@lastpass.com>