endpoints/xml: add endpoints needed for change MPW
There are two calls to the server made when performing a master password
change.  The first retrieves the cipher-text for all of the fields in a
user's vault, along with the encrypted private sharing key and sharing
public keys of any admins (if enterprise).  The second is a submission
of the data newly encrypted with the new key, along with the new login
hash, and, if enterprise super admin recovery is enabled, a copy of the
key encrypted with the admins' public keys.

This change implements all of the xml parsing and http POSTs for
these two steps.

Signed-off-by: Bob Copeland <copeland@lastpass.com>