lpass: implement cmd-mv, for moving sites around
The 'lpass mv' command can be used to move accounts into
different groups, or into and out of shared folders.

One caveat: currently, all items in the account are reencrypted
with new IVs when the blob is saved.  This results in a lot of
needless updates and bogus "password changed" events on the server
side when moving an account to a different folder.  This will be
addressed separately.

Signed-off-by: Bob Copeland <copeland@lastpass.com>