blob: avoid re-encrypting on write-out
Currently, every time we save the blob, we re-encrypt every
account with the current key.  This is not very performant and
unnecessary anyway as we already have encrypted versions of the
plaintext from the last time we fetched the blob.

Additionally, it causes problems when moving sites because the IV
and thus crypttext of every field is updated, resulting in a flood
of 'Site password changed' emails even though the underlying
plaintext remains unchanged.

To fix, only update the crypttext if the plaintext changed, and don't
re-encrypt when saving.

Signed-off-by: Bob Copeland <copeland@lastpass.com>