edit: prohibit SF changes during edit
When a site is edited and the name is changed from a non-shared
folder to a shared folder, or from a shared folder to some other
folder, we need to make an API call to duplicate the site with a
new aid and delete the old one.

At the same time, other changes may have been made to the details
of the site which need to be updated first.  Although we could
potentially support this, it's a bit of extra complexity that is
really best avoided since 'lpass mv' is made for this.  So just
disallow it.

Signed-off-by: Bob Copeland <copeland@lastpass.com>