show: add support for multiple ids
This change allows one to supply multiple ids or search strings as
arguments to lpass show, in order to see all the results in a single
command invocation.

For now, if multiple results are returned, then the output will be
the same as if a search matched multiple items: that is, only ids
will be shown.  This will be addressed in a follow-up patch.

Signed-off-by: Angus Galloway <angustwitter@gmail.com>
[copeland@lastpass.com: reworked to do all the searches up front,
 and modified search routines to handle multiple passes]
Signed-off-by: Bob Copeland <copeland@lastpass.com>