upload-queue: don't exit when first error is encountered
If we get a 4/500 error for some reason during a sync, the
upload program will currently quit at the first opportunity.
This can be annoying if doing a batch of updates, as syncs
halt altogether.  And we might get the error in the first
place because we have a batch and hit some ratelimits.

Introduce a version of http_post_lastpass_v that doesn't
die() in such cases and use it.  In a follow-up patch I'll
tweak the timeouts a bit so we handle this somewhat more
gracefully.

Signed-off-by: Bob Copeland <copeland@lastpass.com>