upload-queue: fix ordering of requests
We generate request filenames by concatenating the unix timestamp
with an increasing ordinal value.

If we have more than 10 requests in a single second (which is
not very hard to achieve) then the generated serial numbers will
be something like (ts)0, ..., (ts)10, (ts)11, (ts+1)0, (ts+1)1 ...

In this case, (ts)10 will sort after (ts+1)0, and so the requests
are sent out of order.  Although still not perfect, we can do much
better by simply zero-padding the ordinal, so do that.

Signed-off-by: Bob Copeland <copeland@lastpass.com>