edit: add missing null check for value
Clang analyzer caught that in some cases we null-check value,
while in this case we do not.  This seems to be a false positive,
as if getline returns successfully it should have allocated a
buffer for us, but in the interest of consistency, treat null
value the same as empty string here.

Signed-off-by: Bob Copeland <copeland@lastpass.com>