blob: implement better support for linked accounts
Linked accounts appear as shared folders but without the Shared-X
naming convention.  In the past, lpass mv would create a non-shared
folder with the same name as the corresponding linked account due
to up-front checks about the naming convention.  Relax these checks
to only enforce them when moving to a non-shared or missing shared
folder; this lets lpass mv do the right thing when the target is
a linked account.

Signed-off-by: Bob Copeland <copeland@lastpass.com>