export: support secure notes
This patch adds export of secure notes to match the exported
records from the website.  In addition, it adds parsing of
the favorites field from the blob so that this field may
be properly exported.

The fullname field is dropped; in a future commit we'll add
back shared folders using the format that the website expects.

Signed-off-by: Justen Walker <justen.walker+github@gmail.com>
[reword, drop group bits, use bool_str()]
Signed-off-by: Bob Copeland <copeland@lastpass.com>