upload-queue: add a log file for debugging sync issues
Users may set the undocumented variable LPASS_LOG_LEVEL=7 in
their environment in order to see debug messages in
~/.lpass/lpass.log.

For now the logging functions are local to upload-queue; as
pretty much everything else is interactive, I don't expect to use
this for much else.  Should we need it, though, basic conditional
logging is supported.

Signed-off-by: Bob Copeland <copeland@lastpass.com>